requirements-core.txt
opencv-python
nltk
pyzbar
deepface
dlib
scipy
textblob
seaborn
textblob_nl
scikit-learn
timm
googletrans==4.0.0-rc1
numpy
torch
torchvision
gradio
pandas
transformers
pillow
ultralytics
matplotlib
pytesseract
vaderSentiment
mediapipe
jupyter
ipython
whisper
tf-keras
pyzbar

